Threads. It's ready. Stamp it!
There's bottlenecks in turning large bigints into a base10 string.
It's substantially faster in base2 (binary) or faster and smaller in base16 (hex).
If I remember right base2 is the fastest.

It's hard to work around it as anywhere you offload it will block.
Which seems to block sending back a response anywhere you do it :)